WIP - fix responsive & add base for comment form
