+++ src/stories/containers/transparency-report/transparency-auditor-comments/comment-form.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import GenericCommentCard from './generic-comment-card';
+
+const CommentForm: React.FC = () => {
+  return (
+    <GenericCommentCard>
+      <h1>CommentForm</h1>
+    </GenericCommentCard>
+  );
+};
+
+export default CommentForm;
