+++ src/stories/containers/transparency-report/transparency-auditor-comments/comment-list.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { ExpenseReportStatus } from '../../../../core/enums/expense-reports-status.enum';
 import AuditorCommentCard from './auditor-comment-card';
 import CUNewExpenseReport from './cu-new-expense-report';
+import CommentForm from './comment-form';
 
 const AuditorCommentList: React.FC = () => {
   return (
@@ -29,6 +30,7 @@ const AuditorCommentList: React.FC = () => {
           'Our September forecast included offsite participation estimates for more people that ended up participating. Equally we have managed to get speaker tickets.\n\n**Updating:**\n- Actual expenses\n- FTE number'
         }
       />
+      <CommentForm />
     </div>
   );
 };
